Fix compile warnings: -Wsign-conversion, complete stage 3
Stage 3 of plans/warning-cleanup.md, finishing the remaining tail
(largest-first, one file/cluster per prior commit; this is the last
batch). Mechanical index/argument casts throughout, plus one real
system-header artifact: LaunchpadIO.cpp's snd_seq_ev_set_sysex() call
expands (it's a macro, <alsa/seqmid.h>) to a plain ~mask-style bit
clear on an unsigned char flags field - GCC attributes the diagnostic
to the macro's expansion site rather than its definition, so the usual
system-header suppression doesn't apply (same class of issue as
RealFFT.h's PocketFFT wrapper) - silenced with a local
#pragma GCC diagnostic push/ignore/pop around just that call.

Clean build: 32 -> 0 warnings. -Wsign-conversion and -Wsign-compare
are both now fully clean.

Verified: clean rebuild (0 errors, 0 warnings), full ctest pass,
--render smoke test.

+++ FileInstrument.cpp
@@ -48,7 +48,7 @@ FileInstrument::openFile() {
   for (int c = 0; c < channels; c++) {
     auto out_buffer = samples_->getChannelData(c);
     for (int i = 0; i < total_frames; i++) {
-      out_buffer[i] = buffer[i * channels + c];
+      out_buffer[i] = buffer[static_cast<size_t>(i * channels + c)];
     }
   }
   return true;
